requirements.txt
streamlit
google-genai
langchain-google-genai
langgraph
langchain
pydantic